requirements.txt
chardet==5.2.0
ImageHash==4.3.2
numpy==2.3.2
pillow==11.3.0
pydub==0.25.1
pysrt==1.1.2
PyWavelets==1.9.0
scipy==1.16.1
webvtt-py==0.5.1
